+++ CHANGELOG.md
@@ -7,6 +7,10 @@
 
 - Update dependencies
 
+### Bug Fix
+
+- Switch players on finger press instead of release ([#6](https://github.com/ldeso/blitz/issues/6))
+
 ## [2.0.6](https://github.com/ldeso/blitz/releases/tag/v2.0.6) – 2025-11-24
 
 This release updates dependencies.